feat: local game mode #97 - fixed found issues env/code duplication/tabs to space
